# server/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'Arya Payroll'
  description: 'This is an API documentation of Arya Payroll'
servers:
  - description: Deploy server
    url: http://192.168.1.119:3002
  - description: LAN Server
    url: http://192.168.1.146:3002
  - description: localmachine server
    url: http://localhost:3002
...
paths:
...
  "/company/update-branch/{companyParentId}/{branchId}":
    post:
      tags:
        - Company Branch
      summary: Update the branch
      operationId: update branch
      parameters:
        - in: path
          name: companyParentId
          required: true
          description: Need parent id to update branch
          schema:
            type: string
        - in: path
          name: branchId
          required: true
          description: Need branch id to update individual
          schema:
            type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  description: Name of the branch
                  type: string
                address:
                  description: Address of the branch
                  type: string
                officialNumber:
                  description: Official Number of the branch
                  type: string
                phone:
                  description: Enter phone number of the branch
                  type: string
                fax:
                  description: Fax number of the branch
                  type: string
                gstNumber:
                  description: GSTNo
                  type: string
                currency:
                  description: Home Currency
                  type: string
                  enum:
                    - USD
                    - INR
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
...
  "/company/delete-branch/{branchId}":
    delete:
      tags:
        - Company Branch
      summary: Delete the branch
      operationId: delete branch
      parameters:
        - in: path
          name: branchId
          required: true
          description: Need branch id to delete branch
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
...
  "/user/add-user-group":
    post:
      tags:
        - User Group
      summary: Add User Group
      operationId: addUserGroup
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  description: User Group Title
                  type: string
                description:
                  description: Description of the User Group
                  type: string
              required:
                - title
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
  "/user/update-user-group/{userGroupId}":
    post:
      tags:
        - User Group
      summary: Update the User Group
      operationId: updateUserGroup
      parameters:
        - in: path
          name: userGroupId
          required: true
          description: Need user group id to update user group
          schema:
            type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  description: User Group Title
                  type: string
                description:
                  description: Description of the User Group
                  type: string
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
...
  "/user/delete-user-group/{userGroupId}":
    delete:
      tags:
        - User Group
      summary: Delete the User Group
      operationId: deleteUserGroup
      parameters:
        - in: path
          name: userGroupId
          required: true
          description: Need user group id to delete user group
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
...